revise SackNotSupportedError
