+++ pkg/networkpath/traceroute/sack/traceroute_sack_nolinux.go
@@ -1,0 +1,24 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2025-present Datadog, Inc.
+
+//go:build !linux
+
+package sack
+
+import (
+	"context"
+	"errors"
+
+	"github.com/DataDog/datadog-agent/pkg/networkpath/traceroute/common"
+)
+
+var errPlatformNotSupported = &SackNotSupportedError{
+	Err: errors.New("SACK traceroute is not supported on this platform"),
+}
+
+// RunSackTraceroute is not supported
+func RunSackTraceroute(ctx context.Context, p Params) (*common.Results, error) {
+	return nil, errPlatformNotSupported
+}
